requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
motor==3.7.0
boto3==1.37.6
python-jose==3.4.0
pyjwt==2.10.1
passlib==1.7.4
bcrypt==4.3.0
types-passlib==1.7.7.20241221
pydantic-settings==2.8.1
email-validator==2.2.0
babel==2.17.0
python-json-logger==3.2.1
openai==1.65.4
sse-starlette==2.2.1
sentry-sdk==2.22.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.1.0
langchain-google-genai>=0.0.8
langgraph>=0.0.30
langgraph-supervisor>=0.0.4
llama-index>=0.10.40
atlassian-python-api>=3.41.3